+++ prover/server/circuits/spp_merge/fixture_test.go
@@ -419,6 +419,7 @@ func buildMergeFixture(t *testing.T, options mergeFixtureOptions) *mergeWitnessF
 		publicInputPreimage = append(
 			publicInputPreimage,
 			userSigningPkHash,
+			userNullifierPk,
 		)
 	case ringFixtureRail:
 		publicInputPreimage = append(